remove helper module
